@@ -23,36 +23,36 @@
     "d3-shape": "1.3.5",
     "d3-time": "^1.0.11",
     "d3-transition": "^1.2.0",
-    "firebase": "5.9.4",
+    "firebase": "^6.0.2",
     "moment": "^2.24.0",
     "particlesjs": "^2.2.3",
     "prop-types": "^15.7.2",
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "react-firebaseui": "^3.1.2",
-    "react-redux": "7.0.2",
+    "react-redux": "^7.0.3",
     "react-router-dom": "5.0.0",
     "redux": "^4.0.1"
   },
   "devDependencies": {
-    "@babel/core": "7.4.3",
-    "@babel/preset-env": "7.4.3",
+    "@babel/core": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.5",
-    "clean-webpack-plugin": "^2.0.1",
+    "clean-webpack-plugin": "^2.0.2",
     "eslint": "5.16.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-loader": "^2.1.2",
-    "eslint-plugin-import": "2.17.1",
+    "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.12.4",
-    "firebase-tools": "6.6.0",
+    "eslint-plugin-react": "^7.13.0",
+    "firebase-tools": "^6.9.1",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "24.7.1",
+    "jest": "^24.8.0",
     "npm-check": "^5.9.0",
-    "webpack": "4.30.0",
-    "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "3.3.1",
+    "webpack": "^4.31.0",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-server": "^3.3.1",
     "webpack-merge": "^4.2.1",
     "webpack-notifier": "^1.7.0"
   }
