@@ -1,9 +1,12 @@
-const firebase = require('firebase/app');
-require('firebase/auth');
-require('firebase/firestore');
+import * as firebase from 'firebase/app';
+
+import 'firebase/auth';
+import 'firebase/firestore';
+import 'firebase/performance';
 
 const config = require('./config.json');
 
 firebase.initializeApp(config);
+firebase.performance();
 
 export default firebase;
